+++ rl_toolkit/utils/__init__.py
@@ -1,1 +1,2 @@
+from .replay_buffer import make_reverb_dataset  # noqa
 from .variable_container import VariableContainer  # noqa