+++ rl_toolkit/utils/variable_container.py
@@ -8,14 +8,14 @@ class VariableContainer:
         # ---
         db_server: str,
         # ---
-        table_name: str,
+        table: str,
         variables: dict,
     ):
-        self._table_name = table_name
+        self._table = table
         self._variables = variables
 
         # Initializes the reverb client
-        self.tf_client = reverb.TFClient(server_address=f"{db_server}:8000")
+        self.tf_client = reverb.TFClient(server_address=db_server)
 
         # variables signature for variable container table
         self.signature = tf.nest.map_structure(
@@ -25,9 +25,7 @@ class VariableContainer:
         self.dtypes = tf.nest.map_structure(lambda spec: spec.dtype, self.signature)
 
     def update_variables(self):
-        sample = self.tf_client.sample(
-            self._table_name, data_dtypes=[self.dtypes]
-        ).data[0]
+        sample = self.tf_client.sample(self._table, data_dtypes=[self.dtypes]).data[0]
         for variable, value in zip(
             tf.nest.flatten(self._variables), tf.nest.flatten(sample)
         ):
@@ -36,7 +34,7 @@ class VariableContainer:
     def push_variables(self):
         self.tf_client.insert(
             data=tf.nest.flatten(self._variables),
-            tables=tf.constant([self._table_name]),
+            tables=tf.constant([self._table]),
             priorities=tf.constant([1.0], dtype=tf.float64),
         )